(#189) Alterando criterios para tinyint

+++ app/model/post.py
@@ -47,10 +47,10 @@ class Post(db.Model):
         db.Integer, nullable=False, primary_key=True, autoincrement=True
     )
     post_date = db.Column(db.Date, default=date.today(), nullable=False)
-    didactic = db.Column(db.Float, nullable=False)
-    metod = db.Column(db.Float, nullable=False)
-    avaliations = db.Column(db.Float, nullable=False)
-    disponibility = db.Column(db.Float, nullable=False)
+    didactic = db.Column(db.Integer, nullable=False)
+    metod = db.Column(db.Integer, nullable=False)
+    avaliations = db.Column(db.Integer, nullable=False)
+    disponibility = db.Column(db.Integer, nullable=False)
     content = db.Column(db.String(480), nullable=False)
     is_anonymous = db.Column(db.Boolean, nullable=False)
